feat: Shop 좋아요 멤버 데이터 타입 변경

+++ bp-app-api/src/main/java/com/beautify_project/bp_app_api/dto/shop/ShopListFindResult.java
@@ -14,7 +14,7 @@ public record ShopListFindResult(
     @JsonInclude(Include.NON_NULL)
     List<String> facilities,
     String rate,
-    Integer likes,
+    Long likes,
     Boolean likePushed,
     String thumbnailLink
 ) {
